+++ lib/apidiesel/handlers/http_request_helper.rb
@@ -16,13 +16,13 @@ module Apidiesel
         http_request      = HTTPI::Request.new(request.url.try(:to_s))
         http_request.body = payload
 
-        if api_config[:http_basic_username] && api_config[:http_basic_password]
-          http_request.auth.basic(api_config[:http_basic_username], api_config[:http_basic_password])
+        if api_config.http_basic_username && api_config.http_basic_password
+          http_request.auth.basic(api_config.http_basic_username, api_config.http_basic_password)
         end
 
-        http_request.auth.ssl.verify_mode = api_config[:ssl_verify_mode] || :peer
-        http_request.open_timeout         = api_config[:timeout] || 30
-        http_request.read_timeout         = api_config[:timeout] || 30
+        http_request.auth.ssl.verify_mode = api_config.ssl_verify_mode
+        http_request.open_timeout         = api_config.request_timeout
+        http_request.read_timeout         = api_config.request_timeout
 
         if block_given?
           http_request = yield http_request
@@ -31,14 +31,14 @@ module Apidiesel
         request.http_request = http_request
 
         begin
-          response = HTTPI.request(request.endpoint.http_method, http_request)
+          response = HTTPI.request(api_config.http_method, http_request)
           request.http_response = response
         rescue => e
           raise RequestError.new(e, request)
         end
 
         if response.error?
-          raise RequestError.new("#{request.endpoint.http_method} #{request.url} returned #{response.code}", request)
+          raise RequestError.new("#{api_config.http_method} #{request.url} returned #{response.code}", request)
         end
 
         request